moved term filter to a file
previously this was set as an environment variable.  One pretty
important thing to note here is that the current spark streaming api
for twitter doesn't include a way to exclude filter terms. For us that
means that two of the terms we had set on the discourse thread here
http://moxie.rtp.rti.org/t/search-terms-keywords/263/11 need to be
addressed somewhere else. Those terms are:
bong AND NOT ("beer bong" OR beerbong OR "bong bong" OR "bing bong")

and
(kush AND NOT Chikki) OR